Modularize things more and work around Julia age range bug by not extending RigidBodyTreeInspector.animate.
