+++ REQUIRE
@@ -5,3 +5,4 @@ DiffEqCallbacks 1.0.0
 RigidBodyTreeInspector 0.3.2
 LoopThrottle 0.0.1
 DocStringExtensions 0.4.1
+Reexport 0.1
